requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
bcrypt
pyjwt
pydantic

